MTL builder, LoggerInfo implemented
