+++ logback-mtl-builder-app/src/main/scala/loggingexperiment/logbackmtl/Main.scala
@@ -3,15 +3,45 @@ package loggingexperiment.logbackmtl
 import java.security.InvalidParameterException
 
 import cats.effect._
+import com.olegpy.meow.monix._
 import io.circe.generic.auto._
 import monix.eval._
 import monix.execution.Scheduler
+import slf4cats._
 
 import scala.language.higherKinds
+import scala.reflect.ClassTag
+
+object MonixLog {
+  def make(
+    logger: org.slf4j.Logger
+  )(taskLocalContext: TaskLocal[Logger.Context]): Logger[Task] = {
+    taskLocalContext.runLocal { implicit ev =>
+      Logger.make(logger)
+    }
+  }
+
+  def make(
+    name: String
+  )(taskLocalContext: TaskLocal[Logger.Context]): Logger[Task] = {
+    taskLocalContext.runLocal { implicit ev =>
+      Logger.make(name)
+    }
+  }
+
+  def make[T](
+    taskLocalContext: TaskLocal[Logger.Context]
+  )(implicit classTag: ClassTag[T]): Logger[Task] = {
+    taskLocalContext.runLocal { implicit ev =>
+      Logger.make
+    }
+  }
+}
 
 object Main extends TaskApp {
 
   final case class A(x: Int, y: String)
+
   final case class B(a: A, b: Boolean)
 
   private val o = B(A(123, "Hello"), b = true)
@@ -24,7 +54,7 @@ object Main extends TaskApp {
   def init(implicit sch: Scheduler): Task[Unit] =
     for {
       mdc <- TaskLocal(Logger.Context.empty)
-      logger = MonixLog.make(mdc)
+      logger = MonixLog.make[Main.type](mdc)
       result <- program(logger)
     } yield result
 
@@ -34,11 +64,10 @@ object Main extends TaskApp {
       _ <- logger
         .context("a", A(1, "x"))
         .context("o", o)
-        .apply
         .info("Hello Monix")
-      _ <- logger.apply.info("Hello MTL", ex)
+      _ <- logger.info("Hello MTL", ex)
       _ <- logger.context("x", 123).context("o", o).use {
-        logger.context("x", 9).apply.info("Hello2 meow")
+        logger.context("x", 9).info("Hello2 meow")
       }
     } yield ()
   }
